+++ src/ShaderPlayground.Core/Compiler.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using ShaderPlayground.Core.Compilers.Dxc;
 using ShaderPlayground.Core.Compilers.Fxc;
 using ShaderPlayground.Core.Compilers.Glslang;
+using ShaderPlayground.Core.Compilers.GlslOptimizer;
 using ShaderPlayground.Core.Compilers.HlslCc;
 using ShaderPlayground.Core.Compilers.Mali;
 using ShaderPlayground.Core.Compilers.Slang;
@@ -29,6 +30,7 @@ namespace ShaderPlayground.Core
             new DxcCompiler(),
             new FxcCompiler(),
             new GlslangCompiler(),
+            new GlslOptimizerCompiler(),
             new HlslCcCompiler(),
             new MaliCompiler(),
             new SlangCompiler(),